Share scale and light-norm buffers across wgpu pipes
