+++ kas-wgpu/src/draw/shaded_square.rs
@@ -5,14 +5,13 @@
 
 //! Simple pipeline for "square" shading
 
-use std::f32;
 use std::mem::size_of;
 use wgpu::util::DeviceExt;
 
 use crate::draw::{Rgb, ShaderManager};
 use kas::cast::Cast;
 use kas::draw::{Colour, Pass};
-use kas::geom::{Quad, Size, Vec2, Vec3};
+use kas::geom::{Quad, Vec2, Vec3};
 
 #[repr(C)]
 #[derive(Clone, Copy, Debug)]
@@ -29,7 +28,6 @@ pub struct Pipeline {
 /// Per-window state
 pub struct Window {
     bind_group: wgpu::BindGroup,
-    scale_buf: wgpu::Buffer,
     passes: Vec<Vec<Vertex>>,
 }
 
@@ -135,21 +133,12 @@ impl Pipeline {
     }
 
     /// Construct per-window state
-    pub fn new_window(&self, device: &wgpu::Device, size: Size, light_norm: [f32; 3]) -> Window {
-        type Scale = [f32; 2];
-        let scale_factor: Scale = [2.0 / size.0 as f32, -2.0 / size.1 as f32];
-        let scale_buf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("SS scale_buf"),
-            contents: bytemuck::cast_slice(&scale_factor),
-            usage: wgpu::BufferUsage::UNIFORM | wgpu::BufferUsage::COPY_DST,
-        });
-
-        let light_norm_buf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("SS light_norm_buf"),
-            contents: bytemuck::cast_slice(&light_norm),
-            usage: wgpu::BufferUsage::UNIFORM | wgpu::BufferUsage::COPY_DST,
-        });
-
+    pub fn new_window(
+        &self,
+        device: &wgpu::Device,
+        scale_buf: &wgpu::Buffer,
+        light_norm_buf: &wgpu::Buffer,
+    ) -> Window {
         let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
             label: Some("SS bind group"),
             layout: &self.bind_group_layout,
@@ -175,7 +164,6 @@ impl Pipeline {
 
         Window {
             bind_group,
-            scale_buf,
             passes: vec![],
         }
     }
@@ -208,11 +196,6 @@ impl Pipeline {
 }
 
 impl Window {
-    pub fn resize(&mut self, queue: &wgpu::Queue, size: Size) {
-        let scale_factor = [2.0 / size.0 as f32, -2.0 / size.1 as f32];
-        queue.write_buffer(&self.scale_buf, 0, bytemuck::cast_slice(&scale_factor));
-    }
-
     /// Add a rectangle to the buffer
     pub fn rect(&mut self, pass: Pass, rect: Quad, col: Colour) {
         let aa = rect.a;
